scripts/build_summary.py
#!/usr/bin/env python3
"""
build_summary.py — Generates the Executive Summary Word document.

Usage:
    python scripts/build_summary.py
"""

import sys, os
sys.path.insert(0, os.path.expanduser("~/.local/lib/python3.12/site-packages"))
sys.path.insert(0, os.path.join(os.path.dirname(__file__), ".."))

from docx import Document
from src import config
from src.utils import ensure_dirs

ensure_dirs(config.OUTPUT_DIR)

doc = Document()

# ── Title ────────────────────────────────────────────────────────────────────────
doc.add_heading("Executive Summary: Expense Classification Pipeline", 0)
doc.add_paragraph(
    "AI/ML/NLP Candidate Test — FY23 Expense Data (268 rows, 208 unique remarks)"
)

# ── 1. Problem Framing ──────────────────────────────────────────────────────────
doc.add_heading("1. Problem Framing", level=1)
doc.add_paragraph(
    'The dataset contains 268 expense records with a free-text "Remarks" field. '
    "The task is to classify each remark into Services, Equipment, or Material. "
    "A critical finding during initial audit: the dataset has NO pre-existing labels. "
    "This makes the task a zero-shot or unsupervised classification problem, not a "
    "standard supervised one."
)
doc.add_paragraph(
    "Before any modelling, we wrote a one-paragraph taxonomy specification defining "
    "each class with boundary criteria tight enough that two independent reviewers "
    "would agree on every borderline case. We also added a fourth class — OTHER — "
    "for accounting journal entries that must never be force-fit into procurement categories."
)

# ── 2. EDA Findings ─────────────────────────────────────────────────────────────
doc.add_heading("2. Key EDA Findings That Changed What We Built", level=1)
p = doc.add_paragraph()
p.add_run("Finding 1: ").bold = True
p.add_run("Year ≡ FY23, Credit ≡ 0, Net ≡ Debit. Three columns are constant/redundant → dropped.\n")
p.add_run("Finding 2: ").bold = True
p.add_run("1 row has a missing Remark (index 124). Routed to OTHER.\n")
p.add_run("Finding 3: ").bold = True
p.add_run("268 rows but only 208 unique remark strings → 59 near-duplicates.\n")
p.add_run("Finding 4: ").bold = True
p.add_run(
    "12 rows are accounting journal entries (provisions, reclasses, CWIP transfers). "
    "A deterministic regex pre-filter routes these to OTHER before the model ever sees them.\n"
)
p.add_run("Finding 5: ").bold = True
p.add_run(
    "Text contains heavy noise — vendor PO prefixes, invoice numbers, date strings. "
    "A normalisation pass strips this before featurisation."
)

# ── 3. Architecture Comparison ───────────────────────────────────────────────────
doc.add_heading("3. Two Architectural Approaches Evaluated", level=1)

p = doc.add_paragraph()
p.add_run("Approach 1 — Zero-Shot Transformer (chosen):\n").bold = True
p.add_run(
    "Uses typeform/distilbert-base-uncased-mnli, a model trained on Natural Language "
    "Inference. Classifies text without any labelled training data. "
    "Advantages: no manual rule bias, handles novel phrasing, easily extensible to new classes. "
    "Disadvantages: computationally heavier, struggles with domain-specific procurement jargon.\n\n"
)
p.add_run("Approach 2 — TF-IDF + LinearSVC with pseudo-labels:\n").bold = True
p.add_run(
    "Generate pseudo-labels via keyword heuristics, then train a lightweight classifier. "
    "Advantages: fast, interpretable feature weights, low infrastructure. "
    "Disadvantages: strongly bounded by rule quality, circular evaluation risk."
)

# ── 4. Pipeline Design ──────────────────────────────────────────────────────────
doc.add_heading("4. Pipeline Design", level=1)
doc.add_paragraph(
    "The pipeline is structured as a modular, config-driven Python package:"
)
table = doc.add_table(rows=6, cols=2, style="Light Shading Accent 1")
table.cell(0, 0).text = "Module"
table.cell(0, 1).text = "Responsibility"
table.cell(1, 0).text = "config.py"
table.cell(1, 1).text = "Single source of truth — all thresholds, patterns, taxonomy. No magic strings elsewhere."
table.cell(2, 0).text = "preprocessing.py"
table.cell(2, 1).text = "Text normalisation (strip PO numbers, dates, invoice codes) + journal pre-filter → OTHER."
table.cell(3, 0).text = "classifier.py"
table.cell(3, 1).text = "Zero-shot classification + confidence-based human review queue export."
table.cell(4, 0).text = "evaluation.py"
table.cell(4, 1).text = "Self-consistency, learnability alarm (TF-IDF+SVC CV), gold-standard eval."
table.cell(5, 0).text = "run_pipeline.py"
table.cell(5, 1).text = "One-command orchestrator. Runs all steps sequentially with structured logging."

# ── 5. Results ───────────────────────────────────────────────────────────────────
doc.add_heading("5. Evaluation Results", level=1)
p = doc.add_paragraph()
p.add_run("Metrics chosen:\n").bold = True
p.add_run("• Macro F1-Score: weights all classes equally regardless of support size.\n")
p.add_run("• Accuracy: raw correctness.\n")
p.add_run("• Confusion matrix: reveals which exact classes are cross-misclassified.\n\n")

p.add_run("Results:\n").bold = True
p.add_run("• Overall accuracy: 47.9%\n")
p.add_run("• Macro F1: 0.588\n")
p.add_run("• OTHER class: F1 = 0.95 (pre-filter works perfectly)\n")
p.add_run("• Self-consistency violations: 0\n")
p.add_run("• Learnability CV F1: 0.778 (alarm fires correctly)\n")
p.add_run("• Human review queue: 128 rows flagged (confidence < 0.55)\n\n")

p.add_run("Key takeaway: ").bold = True
p.add_run(
    "The pipeline infrastructure is sound — pre-filter, normalisation, consistency checks, "
    "learnability alarm, human review queue — but the zero-shot model struggles with "
    "domain-specific procurement terminology. The next step would be active-learning: "
    "have a domain expert correct the 128 lowest-confidence predictions, then fine-tune."
)

# ── 6. Final Distribution ───────────────────────────────────────────────────────
doc.add_heading("6. Final Classification Distribution", level=1)
table2 = doc.add_table(rows=5, cols=2, style="Light Shading Accent 1")
table2.cell(0, 0).text = "Class"
table2.cell(0, 1).text = "Count"
table2.cell(1, 0).text = "Material"
table2.cell(1, 1).text = "112"
table2.cell(2, 0).text = "Equipment"
table2.cell(2, 1).text = "109"
table2.cell(3, 0).text = "Services"
table2.cell(3, 1).text = "35"
table2.cell(4, 0).text = "OTHER"
table2.cell(4, 1).text = "12"

doc.save(config.OUTPUT_SUMMARY)
print(f"Executive_Summary.docx → {config.OUTPUT_SUMMARY}")
